+++ course/src/main/java/in/testpress/course/viewmodels/ProductCategoriesViewModel.kt
@@ -1,0 +1,12 @@
+package `in`.testpress.course.viewmodels
+
+import `in`.testpress.course.repository.ProductCategoriesRepository
+import androidx.lifecycle.ViewModel
+
+class ProductCategoriesViewModel(val repository: `in`.testpress.course.repository.ProductCategoriesRepository): ViewModel() {
+    val items = repository.resourceProductCategories
+
+    fun loadContents() {
+        return repository.loadItems()
+    }
+}
